refactor(vnext): consolidate L2 into src/vnext/gateway/ (#623 step 4)
Replaces the earlier routing/ + transport/ split. At ~10 files total the
two-folder split was premature taxonomy — same layer, same folder. Naming
borrows openclaw's gateway/ for "messages flowing into and out of the
agent": both routing decisions (bindings, mention, session-keys) and
data-processing primitives (dedupe, debounce, ...) live here.

This PR pulls the already-merged dedupe + debounce (step 2) into gateway/
alongside this step's routing modules.

- src/core/{session-keys,bindings,mention}.{ts,test.ts} → src/vnext/gateway/
- src/vnext/transport/{dedupe,debounce}.{ts,test.ts} → src/vnext/gateway/
- discord.ts is the only external importer
- bindings.{ts,test.ts}: ./types.js → ../../core/types.js (types migrates later)
- bindings.test.ts: ../vnext/config.js → ../config.js (sibling)
- 61/61 tests pass
