+++ src/plugins/discord/discord.ts
@@ -21,7 +21,7 @@ import type { ThreadBindingConfig } from "./types.js";
 import type { DefaultAgentManager } from "../../core/agent-manager.js";
 import { userMessage as mkUserMsg, userMessageWithImages as mkUserMsgWithImages } from "../../core/messages.js";
 import type { ContextConfigFile } from "../../vnext/config.js";
-import { shouldRespondToMessage } from "../../core/mention.js";
+import { shouldRespondToMessage } from "../../vnext/gateway/mention.js";
 import { loggers } from "../../vnext/logging/logger.js";
 import { ThreadBindingManager } from "./thread-bindings.js";
 import { consumeRootRun, cancelRunBySessionId, isRootRunActive } from "../../core/agent-run.js";
@@ -31,10 +31,10 @@ import { isSilentReply } from "../../core/silent-reply.js";
 import { extractDiscordMetadata, formatInboundMeta } from "./message-metadata.js";
 import { createReplyResolver, type ReplyToMode } from "./reply-directive.js";
 import type { UsageTracker } from "../../core/usage-tracker.js";
-import { buildSessionKey } from "../../core/session-keys.js";
+import { buildSessionKey } from "../../vnext/gateway/session-keys.js";
 import { ChannelHistoryBuffer, buildHistoryContext } from "../../core/channel-history.js";
-import { DedupeCache } from "../../vnext/transport/dedupe.js";
-import { InboundDebouncer } from "../../vnext/transport/debounce.js";
+import { DedupeCache } from "../../vnext/gateway/dedupe.js";
+import { InboundDebouncer } from "../../vnext/gateway/debounce.js";
 import { SlashCommandHandler } from "../../commands/slash-commands.js";
 
 const log = loggers.discord;
