+++ src/vnext/gateway/debounce.test.ts
@@ -1,4 +1,4 @@
-// src/vnext/transport/debounce.test.ts — Tests for inbound message debouncer
+// src/vnext/gateway/debounce.test.ts — Tests for inbound message debouncer
 
 import { describe, it, expect, vi, afterEach } from "vitest";
 import { InboundDebouncer } from "./debounce.js";
